ai: use integer AI DMA duration scaling instead of float
The AI DMA duration was scaled by a float dma_modifier (ROM_SETTINGS.
aidmamodifier / 100.0), threaded as a float through init_device/init_ai
and multiplied into the duration in get_dma_duration's caller. Floating-
point scaling in a timing-critical path is a determinism hazard (results
can vary with the platform's FP mode/flags) and loses precision once the
duration exceeds 2^24.

Thread the modifier as an integer percentage instead and scale with 64-bit
integer math: duration = (uint64_t)raw * pct / 100. The 64-bit intermediate
avoids overflow at large AI lengths. For pct == 100 (every game except Hey
You Pikachu, which uses 88) this is bit-identical to the previous result --
verified against Ocarina of Time (1201 AI DMAs, zero divergence) -- so it is
a pure determinism/precision hardening with no behavioural change on normal
content.

+++ mupen64plus-core/src/device/rcp/ai/ai_controller.h
@@ -59,7 +59,7 @@ struct ai_controller
     unsigned int samples_format_changed;
     uint32_t last_read;
     uint32_t delayed_carry;
-    float dma_modifier;
+    unsigned int dma_modifier;
 
     struct mi_controller* mi;
     struct ri_controller* ri;
@@ -80,7 +80,7 @@ void init_ai(struct ai_controller* ai,
              struct vi_controller* vi,
              void* aout,
              const struct audio_out_backend_interface* iaout,
-             float dma_modifier);
+             unsigned int dma_modifier);
 
 void poweron_ai(struct ai_controller* ai);
